Initial Release of Genealogy Viewer
The creation and initial release of Genealogy View an application that
specializes in rendering family trees, regardless of family complexity.
This application is targeted not at human family trees but animal trees.
Allowing for capturing of relevant data, and tracking of attributes
across the family trees.

It is capable of rendering multi thousand family trees (granted
readability can get tough) and allow for insight into various aspects of
the population.

Signed-off-by: Nathan Hansen <nate1hansen@gmail.com>